Minor code style changes

+++ Assembler/Library/source/Tokens/Instructions/Token_Instruction_Add.cpp
@@ -40,27 +40,25 @@ std::int32_t Add::assemble(std::vector<std::shared_ptr<Token>> &tokens, Assemble
 
         if (tokens[3]->type() == Token::IMMEDIATE) {
                 if (std::static_pointer_cast<Immediate>(tokens[3])->immediate > 15 ||
-                        std::static_pointer_cast<Immediate>(tokens[3])->immediate < -16) {
+                    std::static_pointer_cast<Immediate>(tokens[3])->immediate < -16) {
                         tokens[3]->expected("5 bit immediate value");
                         return -1;
                 }
         }
 
         assembled.push_back(
-                static_cast<std::uint16_t> (0x1000 |
-                        static_cast<std::uint16_t >(
-                                ((std::static_pointer_cast<Register>(tokens[1])->reg & 7) << 9) |
-                                        ((std::static_pointer_cast<Register>(tokens[2])->reg) & 0x7) << 6)
+                static_cast<std::uint16_t>(0x1000 |
+                                           ((std::static_pointer_cast<Register>(tokens[1])->reg & 7) << 9) |
+                                           ((std::static_pointer_cast<Register>(tokens[2])->reg) & 0x7) << 6
                 )
         );
 
         if (tokens[3]->type() == Token::REGISTER) {
                 assembled.front() |= (std::static_pointer_cast<Register>(tokens[3])->reg & 0x7);
         } else {
-                assembled.front() |= (0x20 |
-                        static_cast<std::uint16_t>(
-                                std::static_pointer_cast<Immediate>(tokens[3])->immediate & 0x1F)
-                );
+                assembled.front() |= 0x20 |
+                                     static_cast<std::uint16_t>(
+                                             std::static_pointer_cast<Immediate>(tokens[3])->immediate & 0x1F);
         }
 
         return 1;
